Enable zero-config swagger in vanilla JAX-RS app.
Motivation:
This ensures that swagger can be used with zero configuration as long as
the application's class files fall somewhere outside of the
org.wildfly.swarm package space.

Result:
Now we have 2 swagger examples. One in jaxrs/swagger which shows how to
use the SwaggerFraction explicitly, enabling user-level configuration of
the swagger options. The other is in jaxrs/jaxrs-war which illustrates
how to just use swagger out of the box with no real changes other than
the required swagger annotations on your JAX-RS resources.