Remove fraction dependencies for war projects where appropriate
Motivation:
Since the maven plugin can now detect fractions, we can simplify the WAR
examples to take advantage of that.

Note that detection can't detect a JSF app if it has no classes that
reference JSF packages, since the detection is currently all package
based, so we don't convert that example.

Modifications:
* remove fraction dependencies
* add in the required ee spec deps to allow the projects to compile

Result:
We now have examples of using auto-detection.