SWARM-806: JPA+JAX-RS+CDI+JTA examples have wrong assertions (#103)
Motivation
----------
Both the JPA+JAX-RS+CDI+JTA examples have the following exceptions
in the log:

    Could not find MessageBodyWriter for response object of type:
    java.lang.Integer of media type: text/html

yet the tests don't fail. This is because the FEST Assert API
is used wrongly and no assertions are actually made.

Modifications
-------------
Fix usage of the FEST Assert API so that asserts are actually made.
This itself makes the tests fail, but that is easy to fix: just
make the int-returning JAX-RS methods @Produces("text/plain"),
so that java.lang.Integer marshaller is found.

Result
------
More test coverage, less exceptions in the log, tests still pass.