SWARM-851: WAR custom main test (#111)
* add log name to LogAssert assertion message

Motivation
----------
If a log-related assertion fails, the assertion message doesn't show
in which log file an expectation wasn't met.

Modifications
-------------
Change AbstractIntegrationTest to pass the log file name to the Log
class. Change LogAssert to use that in the assertion message.

Result
------
More details in LogAssert assertion messages.

* SWARM-851: add a test that a custom main method in a WAR project is actually executed

Motivation
----------
The `jaxrs-war-main` example demonstrates that a WAR project can have
a custom `main` method, but it isn't actually tested. The existing
tests in the example wouldn't reveal a situation when the custom
`main` method would be skipped.

Modifications
-------------
The custom `main` method in the `jaxrs-war-main` example now prints
a message to stdout. The test asserts that the message is present
in the log file.

Result
------
More test coverage.