THORN-1531: Following instructions at wildfly-swarm-example/ribbon-consul doesn't yield desirable result.
Motivation
----------
If you follow the instructions on the ribbon consul examples it won't work because topology connector will try to advertise 0.0.0.0: to consul - which is invalid. You need to supply a valid ip address for swarm.bind.address - that should be mentioned in the readme.
Also the examples still use Main class which is deprecated..

Modifications
-------------
Changed the readme to reflect the necessary command line parameter.

Result
------
Examples will be better understandable and runnable by newcomers.