THORN-2168: the "uberjar" profile in Examples no longer works
Motivation
----------
As part of the rename from WildFly Swarm to Thorntail, the `uberjar`
profile in Examples was changed to set the `thorntail.useUberJar`
property instead of `wildfly-swarm.useUberJar`. That change was
wrong. To keep backwards compabitility, we still use the old name
`wildfly-swarm.useUberJar`.

Modifications
-------------
Restore the original property name.

Result
------
The `uberjar` profile works again.